refactor: move useFetchGroupsByCategory into useGroups
