+++ app/api/groups/byCategory/[category]/route.ts
